+++ crates/wdl-engine/src/inputs.rs
@@ -6,6 +6,7 @@ use std::collections::HashSet;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::Path;
+use std::pin::Pin;
 
 use anyhow::Context;
 use anyhow::Result;
@@ -20,6 +21,7 @@ use wdl_analysis::document::Input;
 use wdl_analysis::document::Task;
 use wdl_analysis::document::Workflow;
 use wdl_analysis::types::CallKind;
+use wdl_analysis::types::CallType;
 use wdl_analysis::types::Coercible as _;
 use wdl_analysis::types::Optional;
 use wdl_analysis::types::PrimitiveType;
@@ -59,6 +61,9 @@ fn check_input_type(document: &Document, name: &str, input: &Input, value: &Valu
 
     Ok(())
 }
+
+/// The maximum number of nested calls to recurse before erroring.
+const MAX_NESTED_INPUT_DEPTH: usize = 5;
 
 /// Represents inputs to a task.
 #[derive(Default, Debug, Clone)]
@@ -129,14 +134,28 @@ impl TaskInputs {
     pub async fn join_paths<'a>(
         &mut self,
         task: &Task,
-        path: impl Fn(&str) -> Result<&'a EvaluationPath>,
+        resolver: &(dyn Fn(&str) -> Result<&'a EvaluationPath> + Send + Sync),
+    ) -> Result<()> {
+        self.join_paths_with_prefix(task, task.name(), resolver)
+            .await
+    }
+
+    /// The logic for [`Self::join_paths()`].
+    ///
+    /// This is split out to allow specifying the `prefix` for input paths in
+    /// nested calls.
+    async fn join_paths_with_prefix<'a>(
+        &mut self,
+        task: &Task,
+        prefix: &str,
+        resolver: &(dyn Fn(&str) -> Result<&'a EvaluationPath> + Send + Sync),
     ) -> Result<()> {
         for (name, value) in self.inputs.iter_mut() {
             let Some(ty) = task.inputs().get(name).map(|input| input.ty().clone()) else {
                 bail!("could not find an expected type for input {name}");
             };
 
-            let base_dir = path(name)?;
+            let base_dir = resolver(&format!("{prefix}.{name}"))?;
 
             if let Ok(v) = value.coerce(None, &ty) {
                 *value = v
@@ -431,15 +450,32 @@ impl WorkflowInputs {
     /// expected types.
     pub async fn join_paths<'a>(
         &mut self,
+        document: &Document,
         workflow: &Workflow,
-        path: impl Fn(&str) -> Result<&'a EvaluationPath>,
+        resolver: &(dyn Fn(&str) -> Result<&'a EvaluationPath> + Send + Sync),
+    ) -> Result<()> {
+        self.join_paths_with_prefix(document, workflow, workflow.name(), 0, resolver)
+            .await
+    }
+
+    /// The logic for [`Self::join_paths()`].
+    ///
+    /// This is split out to allow specifying the `prefix` for input paths in
+    /// nested calls.
+    async fn join_paths_with_prefix<'a>(
+        &mut self,
+        document: &Document,
+        workflow: &Workflow,
+        prefix: &str,
+        depth: usize,
+        resolver: &(dyn Fn(&str) -> Result<&'a EvaluationPath> + Send + Sync),
     ) -> Result<()> {
         for (name, value) in self.inputs.iter_mut() {
             let Some(ty) = workflow.inputs().get(name).map(|input| input.ty().clone()) else {
                 bail!("could not find an expected type for input {name}");
             };
 
-            let base_dir = path(name)?;
+            let base_dir = resolver(&format!("{prefix}.{name}"))?;
 
             if let Ok(v) = value.coerce(None, &ty) {
                 *value = v
@@ -447,6 +483,38 @@ impl WorkflowInputs {
                     .await?;
             }
         }
+
+        if !workflow.allows_nested_inputs() {
+            return Ok(());
+        }
+
+        let workflow_calls = workflow.calls();
+        for (call_name, nested_inputs) in self.calls.iter_mut() {
+            let Some(call) = workflow_calls.get(call_name) else {
+                bail!("could not find call {call_name} in workflow");
+            };
+
+            let target_doc = call
+                .namespace()
+                .map(|ns| {
+                    document
+                        .namespace(ns)
+                        .expect("namespace should be present")
+                        .document()
+                })
+                .unwrap_or(document);
+
+            nested_inputs
+                .join_paths(
+                    target_doc,
+                    call,
+                    &format!("{prefix}.{call_name}"),
+                    depth + 1,
+                    resolver,
+                )
+                .await?;
+        }
+
         Ok(())
     }
 
@@ -847,6 +915,50 @@ impl Inputs {
         })?);
 
         Self::parse_json_object(document, object)
+    }
+
+    /// Used to recursively join paths.
+    ///
+    /// See [`TaskInputs::join_paths()`] and [`WorkflowInputs::join_paths()`].
+    fn join_paths<'a, 'b>(
+        &'a mut self,
+        document: &'a Document,
+        call: &'a CallType,
+        prefix: &'a str,
+        depth: usize,
+        resolver: &'a (dyn Fn(&str) -> Result<&'b EvaluationPath> + Send + Sync),
+    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'a>> {
+        Box::pin(async move {
+            if depth > MAX_NESTED_INPUT_DEPTH {
+                bail!(
+                    "maximum nested input depth of {MAX_NESTED_INPUT_DEPTH} exceeded at `{prefix}`"
+                );
+            }
+
+            match self {
+                Inputs::Task(task_inputs) => {
+                    let task = document
+                        .task_by_name(call.name())
+                        .expect("task should be present");
+
+                    task_inputs
+                        .join_paths_with_prefix(task, prefix, resolver)
+                        .await
+                }
+                Inputs::Workflow(workflow_inputs) => {
+                    let workflow = document.workflow().expect("should have a workflow");
+                    assert_eq!(
+                        workflow.name(),
+                        call.name(),
+                        "call name does not match workflow name"
+                    );
+
+                    workflow_inputs
+                        .join_paths_with_prefix(document, workflow, prefix, depth, resolver)
+                        .await
+                }
+            }
+        })
     }
 
     /// Determines if the inputs are empty.